+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.3.0]
+
+* pagination works now good, load 20 more photos each paging
+
+## [0.2.0]
+
+* updated example because the package more flexible now
+* fixed bug of logging in every time, added wrap auth of Instagram 
+to remember past login
+
 ## [0.1.0]
 
 * added Example